refactor: componentsディレクトリの整理
