+++ src/components/organisms/FocusedMarker/focusedMarkerIcon.ts
