+++ src/components/organisms/Markers/index.tsx
@@ -2,12 +2,17 @@ import { useState, useEffect, useRef, createRef, RefObject } from "react";
 import { Marker, useMap } from "react-leaflet";
 import { Marker as MarkerType } from "leaflet";
 import { nanoid } from "nanoid";
-import { RoutePoint, Route, FocusedMarkerInfo, DrawingMode } from "../../types";
+import {
+  RoutePoint,
+  Route,
+  FocusedMarkerInfo,
+  DrawingMode,
+} from "../../../types";
 import { MarkerIcon, GoalMarkerIcon, StartMarkerIcon } from "./markerIcon";
 import {
   routeReducerAction,
   routeAsyncAction,
-} from "../../reducers/routeReducer";
+} from "../../../reducers/routeReducer";
 
 type MakersProps = {
   route: Route;
