+++ src/components/organisms/RouteEditController/index.tsx
@@ -4,12 +4,12 @@ import L from "leaflet";
 import {
   routeReducerAction,
   routeAsyncAction,
-} from "../../reducers/routeReducer";
+} from "../../../reducers/routeReducer";
 import EditableNameDisplay from "../EditableNameDisplay";
 import ElevationGraph from "../ElevationGraph";
-import { config } from "../../config";
-import { Route, FocusedMarkerInfo, DrawingMode } from "../../types";
-import { meters2kilometers } from "../../utils";
+import { config } from "../../../config";
+import { Route, FocusedMarkerInfo, DrawingMode } from "../../../types";
+import { meters2kilometers } from "../../../utils";
 import {
   Button,
   RadioGroup,
@@ -18,7 +18,7 @@ import {
   FormLabel,
   Typography,
 } from "@mui/material";
-import { pagePaths } from "../../consts/uriComponents";
+import { pagePaths } from "../../../consts/uriComponents";
 
 type RouteEditControllerProps = {
   isInsideMap: boolean;
