+++ src/components/pages/RouteIndex/index.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import { getRoutes, postRoutes, deleteRoute } from "../../api/routes";
-import { Route } from "../../types";
-import SignInRequiredTemplate from "../../components/SignInRequiredTemplate";
-import { pagePaths } from "../../consts/uriComponents";
+import { getRoutes, postRoutes, deleteRoute } from "../../../api/routes";
+import { Route } from "../../../types";
+import SignInRequiredTemplate from "../../organisms/SignInRequiredTemplate";
+import { pagePaths } from "../../../consts/uriComponents";
 
 function RouteIndex() {
   const [inputValue, setInputValue] = useState<string>("");
